+++ src/Sim.hpp
@@ -12,36 +12,36 @@
 class Sim {
 	public:
 		Sim();				// default constructor
-		Sim(double ori[3]); // constructor
+		Sim(float ori[3]); // constructor
 
-		double* getJointAngles();
-		double* getEEPose();
-		void moveByJointAngles(double jointAngles[3], double duration);
+		float* getJointAngles();
+		float* getEEPose();
+		void moveByJointAngles(float jointAngles[3], float duration);
 
 		KinematicsModel _km;// kinematics model
 		Trajectory _tjt;    // trajectory model
 
-		double _initJA[NUM_OF_JOINTS];
-		double _currentJA[NUM_OF_JOINTS];
+		float _initJA[NUM_OF_JOINTS];
+		float _currentJA[NUM_OF_JOINTS];
 
 		#ifdef RENDER
 		matrix_t** _actions;	// denormalized actions from rl algo
 		#endif
 		int _numOfActions;
 
-		double _target[CART_DIM];
-		double _obj[CART_DIM];
-		double _initObj[CART_DIM];
+		float _target[CART_DIM];
+		float _obj[CART_DIM];
+		float _initObj[CART_DIM];
 		bool _hasObj;
 		bool _eeState;
-		// double* _target;
+		// float* _target;
 		int _numOfSteps;
 
 	private:
 		
 
-		// double _origin[3];
-		// double _timeStep;
+		// float _origin[3];
+		// float _timeStep;
 };
 
 extern "C" matrix_t* function_call_test();
