+++ src/model/InvKM.h
@@ -8,25 +8,25 @@ extern "C" {
 #include "FwdKM.h"
 
 typedef struct _threeDOFsInv {
-	double l1, l2, l3;	// lengths of the three links
-	double baseHeight;
+	float l1, l2, l3;	// lengths of the three links
+	float baseHeight;
 
 	// /* store the coordinates of two fixed-joints */
-	// double baseCoord[CART_COORD_DIM];
-	// double shoulderCoord[CART_COORD_DIM];
+	// float baseCoord[CART_COORD_DIM];
+	// float shoulderCoord[CART_COORD_DIM];
 
 	/* four angles between links initially, where a2 is fixed */
-	double initA1, initA2, initA3, initA4;
+	float initA1, initA2, initA3, initA4;
 } threeDOFsInv;
 
 
 int initInvKM();
 
-int getAllPoseByJnts(const double jntArray[JNT_NUMBER],
-						double allPoss[JNT_NUMBER][POSE_FRAME_DIM]);
+int getAllPoseByJnts(const float jntArray[JNT_NUMBER],
+						float allPoss[JNT_NUMBER][POSE_FRAME_DIM]);
 
-int getJntsByMagnetPos(const double eePos[POSE_FRAME_DIM],
-						double jntArray[JNT_NUMBER]);
+int getJntsByMagnetPos(const float eePos[POSE_FRAME_DIM],
+						float jntArray[JNT_NUMBER]);
 
 int finishInvKM();
